+++ CHANGELOG.md
@@ -9,6 +9,7 @@ _Tested up to WordPress 5.0_
 
 - Make sure the file exists when trying to get its size.
 - Make sure a directory still exists on the server before deleting it.
+- Add a fallback to wpApiSettings in case Gutenberg is active.
 
 
 ---